+++ apps/web/app/layout.tsx
@@ -7,6 +7,7 @@ import GoogleOneTap from '@/components/GoogleOneTap'
 import ServiceWorkerRegistrar from '@/components/ServiceWorkerRegistrar'
 import CronTrigger from '@/components/CronTrigger'
 import NetworkStatus from '@/components/NetworkStatus'
+import { AdSenseProvider } from '@/components/ads/AdSenseProvider'
 import { createAdminClient } from '@/lib/utils/supabase-admin'
 import { unstable_cache } from 'next/cache'
 
@@ -77,9 +78,11 @@ export default async function RootLayout({
                 <ServiceWorkerRegistrar />
                 <CronTrigger />
                 <UserProvider>
-                    <NetworkStatus />
-                    <GoogleOneTap />
-                    {children}
+                    <AdSenseProvider>
+                        <NetworkStatus />
+                        <GoogleOneTap />
+                        {children}
+                    </AdSenseProvider>
                 </UserProvider>
             </body>
         </html>